Feature/allow yaml config (#40)
* rename generate_cli to generate

* rename start_* to start

* correct multiply by -1

* correct if not else

* use message as ref

* use is_ok

* use negation

* correct result not used

* rename TaskmasterError to error::Taskmaster

* remove unused + use value as ref

* remove useless test

* remove var

* simplify cond

* duplicate use of clone

* use value as ref

* use .values()

* remove redundant else block

* remove redundant closure

* correct history warning

* add editor config

* remove carriage return

* add subject

* add stopdelay + stopsignal

* correct typo on readme

* add default

* add minimal configs

* correct warning

* add task

* use default

* remove cargo run

* correct typo

* move config to folder configs

* edit display string

* format

* add retry conf

* add success delay

* add expected_exit_codes

* add relaunch stratedy

* add config with relaunch strategy

* add mod relaunch

* impl Display trait for relaunch

* add restart field

* add env default value + use static array

* add env

* add example config

* correct configs

* add deps for yaml

* add minimal config example

* check type of file by extension

* rename parse to parsetoml

* add error for yaml parsing

* rename minimal.yml to example.yml

* the maximal size for an octal number for mode is great than 16 bits

* edit example with umask in octal

+++ src/server/task.rs
@@ -21,7 +21,7 @@ pub struct Task {
     cmd: Vec<String>,
     numprocess: u32,
     autostart: bool,
-    umask: u16,
+    umask: u32,
     workingdir: PathBuf,
 
     stdout: PathBuf,
